+++ tests/ConnTest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace KatrinaTest;
+
+use PHPUnit\Framework\TestCase;
+
+class ConnTest extends TestCase
+{
+    public function testInsert()
+    {
+        dd(ORMTest::connection('pgsql')::select()->get());
+    }
+}