+++ tests/ORMTest.php
@@ -10,9 +10,11 @@ class ORMTest extends Katrina
 {
     protected ?string $table = "users";
     //protected ?string $id = "id_orm";
-    protected bool $timestamp = false;
+    protected bool $timestamp = true;
 
     //protected ?bool $cache = true;
+    protected string $created_at = 'created_date';
+    protected string $updated_at = 'updated_date';
 
     public function create()
     {
@@ -21,7 +23,7 @@ class ORMTest extends Katrina
             ->varchar("name", 20)->notNull()
             ->varchar("email", 100)->notNull()
             #->constraint("dev_cons_fk")->foreign("id_usu")->references("usuarios", "idUsu")->onDelete('cascade')
-            #->createdUpdateAt()
+            ->createdUpdatedAt()
             ->closeTable();
 
         return $res;